+++ packages/core/ios/RNSentryReplayBreadcrumbConverter.m
@@ -35,8 +35,8 @@
         return [self convertTouch:breadcrumb];
     }
 
-    if ([breadcrumb.category isEqualToString:@"ui.frustration"]) {
-        return [self convertFrustration:breadcrumb];
+    if ([breadcrumb.category isEqualToString:@"ui.multiClick"]) {
+        return [self convertMultiClick:breadcrumb];
     }
 
     if ([breadcrumb.category isEqualToString:@"navigation"]) {
@@ -79,7 +79,7 @@
                                                                   data:breadcrumb.data];
 }
 
-- (id<SentryRRWebEvent> _Nullable)convertFrustration:(SentryBreadcrumb *_Nonnull)breadcrumb
+- (id<SentryRRWebEvent> _Nullable)convertMultiClick:(SentryBreadcrumb *_Nonnull)breadcrumb
 {
     if (breadcrumb.data == nil) {
         return nil;
@@ -89,7 +89,7 @@
     NSString *message = [RNSentryReplayBreadcrumbConverter getTouchPathMessageFrom:path];
 
     return [SentrySessionReplayHybridSDK createBreadcrumbwithTimestamp:breadcrumb.timestamp
-                                                              category:@"ui.frustration"
+                                                              category:@"ui.multiClick"
                                                                message:message
                                                                  level:breadcrumb.level
                                                                   data:breadcrumb.data];